Implemented provision component with title
