+++ dashboard/src/main/home/onboarding/Onboarding.tsx
@@ -1,5 +1,6 @@
 import React, { useContext, useEffect } from "react";
 import { Context } from "shared/Context";
+import styled from "styled-components";
 import { actions } from "./OnboardingState";
 import Routes from "./Routes";
 
@@ -10,7 +11,18 @@ const Onboarding = () => {
     actions.initFromGlobalContext(context);
   }, [context]);
 
-  return <Routes />;
+  return (
+    <StyledOnboarding>
+      <Routes />
+    </StyledOnboarding>
+  );
 };
 
 export default Onboarding;
+
+const StyledOnboarding = styled.div`
+  width: calc(90% - 130px);
+  min-width: 300px;
+  position: relative;
+  margin-top: calc(50vh - 340px);
+`;
